fix: formatting, clarify names
