+++ crates/starknet-devnet/benches/mint_bench.rs
@@ -8,12 +8,12 @@ pub mod common;
 static DUMMY_ADDRESS: u128 = 1;
 static DUMMY_AMOUNT: u128 = 1;
 
-async fn mint_iter(f: &str) {
+async fn mint_iter(capacity: &str) {
     let devnet = BackgroundDevnet::spawn_with_additional_args(&["--state-archive-capacity", f])
         .await
         .expect("Could not start Devnet");
 
-    for _n in 1..=5000 {
+    for _n in 1..=100 {
         devnet.mint(DUMMY_ADDRESS, DUMMY_AMOUNT).await;
     }
 }
@@ -22,8 +22,9 @@ fn bench_memory(c: &mut Criterion) {
     let rt = Runtime::new().unwrap();
     let mut group = c.benchmark_group("Mint");
     group.significance_level(0.1).sample_size(10);
-    group.bench_function("full", |b| b.to_async(&rt).iter(|| black_box(mint_iter("full"))));
-    group.bench_function("none", |b| b.to_async(&rt).iter(|| black_box(mint_iter("none"))));
+    for i in ["full", "none"].iter() {
+        group.bench_function(*i, |b| b.to_async(&rt).iter(|| black_box(mint_iter(*i))));
+    }
 
     group.finish();
 }
